Add project scaffolding (#1)
`.github/workflows/main.yaml`:
- Print version to make it easier to update `shell.nix` with a checksum
that is presumably fast in CI. Note that adding this step reduces the
time it takes to install the dependencies by about the same time it
takes this step to run (33s + 35s is approximately 67s).
- Use `check` because it is easier to replicate and still fast enough
that timing individual steps is not helpful. s: Reinstate print version
step