chore: Install `mkhelp` in dev env (#3)
`.github/workflows/main.yaml`:
- Add a step to verify that `mkhelp` is available in the development
environment to convince myself that it works and that the cache is used
as expected. The cost of keeping this in to ensure that the development
environment continues to function should be negligible since the program
can be, and is, fetched from cache.
- `y832li5g1aalxmsk9sbp6lj6j5vbd5i5-mkhelp-0.2.3` was uploaded to the
nix cache manually. I'd like to automate this eventually.

`pkgs/mkhelp/default.nix`:
- Inspired by nixpkgs's rust section documentation [^1].
- The meta section appears to be optional, but is kept because recipes I
have looked at consistently include it.

[^1]:
https://github.com/NixOS/nixpkgs/blob/master/doc/languages-frameworks/rust.section.md#buildrustpackage-compiling-rust-applications-with-cargo-compiling-rust-applications-with-cargo