feat: Add discover devices command (#2)
This could easily be accomplished with some command line utilities, but
putting it in Rust code:
- Saves me from having to learn platform specific commands.
- Saves me from having to look up the names of the services to browse
for.
- Facilitates integration with other programs.

`crates/rs4a/src/commands/discover_devices.rs`
---

Use `zeroconf` instead of `async_zeroconf` because the latter did not
compile when I tried it on Linux.

`shell.nix`
---

Add `avahi` because without it compilation fails:
> wrapper.h:1:10: fatal error: 'avahi-client/client.h' file not found

Set `LIBCLANG_PATH` because without it compilation fails:
> Unable to find libclang: "couldn't find any valid shared libraries
matching: ['libclang.so', 'libclang-*.so', 'libclang.so.*',
'libclang-*.so.*'], set the `LIBCLANG_PATH` environment variable to a
path where one of these files can be found (invalid: [])"