chore: Use mDNS directly to discover devices (#10)
This makes it easier to use the tool from inside a container since it
does not require a running avahi daemon, which is tricky and probably a
bad idea.

It should be noted that the command still only works from inside a
container if the host is Linux and the container is started with the
host network driver.