s: fix env?
The `check_lint` target was failing in CI:

> error: unexpected `cfg` condition name: `test`

This is a warning that is promoted to an error because we use `-D warnings`.
The minimal example reproducing the warning is `nix-shell --pure --run 'cargo clippy --all-targets'.

As far as I can tell the issue is that if the cargo bin directory is not on the path, then it is prepended.
Any of these make the error go away:
- Appending it to the path manually, as is done in this commit.
- Removing `~/.cargo/bin`.
- Setting `CARGO_HOME` to another directory.