feat: Add `export-loans` command (#12)
This is intended to make it easier to use the Virtual Loan Tool [^1]
with programs that infer what device to interact with from environment
variables, such `cargo-acap-sdk`.

`shell.nix` is updated because clippy fails on the new integration test:

> error: unexpected `cfg` condition name: `test`

This is a warning that is promoted to an error because we use `-D
warnings`. It can be reproduced like: `nix-shell --pure --run 'cargo
clippy --all-targets`; removing either `--pure` or `--all-targets- makes
the error go away.

I still don't fully understand what is happening and I'm not entirely
happy with the solution. As far as I can tell the issue is that if the
cargo bin directory is not on the path, then it is prepended; any of
these changes makes the error go away:
- Appending it to the path manually, as is done in this commit.
- Removing the `~/.cargo/bin` directory.
- Setting `CARGO_HOME` to another directory.