chore: Add snapshot test for device-inventory
I find that it slow to write tests in rust and that it is tiresome to understand how they map onto actions that I as a human would take. But I also think that maintaining elaborate bash scripts is error prone and tedious. Hopefully I can get the best of both worlds like this.