feat: Add device-inventory program (#13)
This replaces the export command.

The `--offline` option is added because I like being able to control
side effects and because it makes automatic testing easier. It would not
be needed if commands such as list and export didn't automatically
import loans from the virtual loan tool. I don't think this was fully
thought through and that I was thinking it would reduce the number of
commands issued in the terminal to just `export`. This is not true when
there are multiple devices involved, as we must also list them. It is
not even true when there is a single device involved since we must log
in; though this is done only rarely, we likely also want to call some
initialization script when first using a new device.