s: misc fixes
Communicate inventory path via environment variable to avoid having it show up in the output. This is important because it is different on every run. And since `device-inventory help` prints the environment variables, we treat is specially to avoid that happening.

Communicate offline flag via environment variable to be consistent.

Use `set -x` instead of looping because it is easier.