chore: Add snapshot test for device-inventory (#14)
I find that it slow to write tests in rust and that it is tiresome to
understand how they map onto actions that I as a human would take. But I
also think that maintaining elaborate bash scripts is error prone and
tedious. Hopefully I can get the best of both worlds like this.

Communicate inventory path via environment variable to avoid having it
show up in the output. This is important because it is different on
every run. And since `device-inventory help` prints the environment
variables, we treat is specially to avoid that happening.

Communicate offline flag via environment variable to be consistent.

Use `set -x` instead of looping because it is easier.