chore: Rename `rs4a` binary (#18)
The original intent was to make a single program with many sub-
commands, but with `device-inventory` it made sense to group some sub-
commands and adding another level of nesting inside the program would
make the CLI more complicated to implement and navigate, while at the
same time requiring the user to type more with less support from tab-
completion.

Now, having to type both `rs4a` and `discover-devices` is annoying.

Furthermore, I think using a tools that doesn't do a lot of things I
don't need is a nicer experience. The other extreme, having to use many
tools to accomplish any one task, isn't nice either, and I believe that
having one level of subcommands strikes I good balance.

---

`crates/device-finder/src/main.rs`:
- Expose the `DiscoverDevicesCommand` as the CLI entry-point without
moving or renaming it because it is a smaller change, it is still easy
to understand what is going on, it makes it easier to add more commands
in the future, and it doesn't make a difference to the user.