chore: Split snapshot tests into docs and smoke-test (#19)
Prompted by me feeling like it doesn't make sense that the
`device-finder` test is called a smoke test. In theory it also makes
reviewing snapshot changes easier since the name of the file primes the
reviewer for what kind of changes they should look for.

`Makefile`:
- Replace specific targets with a generic to have fewer targets to
maintain. The only drawback is that detecting which binary to build
becomes more difficult, but since rebuilding them is essentially a no-op
thanks to Cargo's cache, this seems like a good trade-off.