feat: Add completions command to all binaries (#20)
Because having tab completions makes programs easier and more enjoyable
to use.

`crates/bin-utils/src/completions_command.rs`:
- The command is factored out to a shared crate straight away because my
experience from previous projects is that it is easy for CLIs to start
diverging and I think that a set of tools is easier to use if their
interface follow similar patterns.

`crates/device-finder/src/main.rs`:
- Make the completions command a separate command on the same level as
`discover-devices` for consistency. This also enables us to have the
`help` command, which further improves consistency.
- Since there is only one _real_ command in this program, this command
is run by default even though this violates my principle that a program
should do as little as possible, and especially have no side effects, by
default. This is done because I expect the default command, which is
simple to use, is what the user will want virtually every time.