chore: Add installation section to readme (#22)
Since the tools target ACAP developers in general, not only those using
Rust, it may not be immediately obvious how to install the tools. It
also may not be convenient to use Cargo, but in that case I don't know
what would be convenient, hence the suggestion to open an issue.