feat: Add client builder (#24)
The main purpose is to make it easier to construct a client when:
- Using local VAPIX as an ACAP app.
- Connecting to a device in development, e.g. when running an ACAP app
on host or when running test suites for this library.

---

`crates/vapix/src/client.rs`:
- Expose the internal builder using `with_inner` to allow users to
customize the behavior of the reqwest client in ways that are not yet
fully supported. This is particularly useful for allowing invalid certs,
which I'm not sure if or how to expose. This method also keeps the
exposure of the reqwest dependency small, paving the road for supporting
other libraries.
- Expose `basic_authentication` even though this could be accomplished
using `with_inner` because it is a common configuration that I have no
qualms about supporting. Furthermore, if support for digest
authentication is added in the future, then customizing the reqwest
client will not be enough.
- The `new` and `build_with_scheme` methods take arguments as these are,
and will likely remain, mandatory arguments and it is convenient to
ensure their presence with the type system.
- One nice feature of the new builder is that most of the setup happens
in synchronous code. This makes it easier to build a client when a
program starts, possibly before the async runtime is started, and be
confident that it happens quickly.