feat!: Build all XML for `add_action_configuration` (#35)
Crafting XML that will be accepted by the server is error prone, mostly
because it is easy to get the namespaces wrong, so relieving the user of
this responsibility is helpful.

Because the problem is typically namespaces, this API is not
particularly valuable, but I start with it because it seems relatively
easy to cover all the variations.

---

`crates/vapix/src/services/action1.rs`:
- Change the function signature to take a template token because this is
probably a mandatory argument in all requests and I want to reduce the
number of invalid requests that can be constructed.

`crates/vapix/tests/serde.rs`:
- Add snapshot testing because this is a cheaper way of catching
regressions than sending requests to a real device, even though the
external validity is lower.
- Replace the `add_action_configuration` example because the existing
request probably isn't valid and doesn't exercise much of the changed
code.