Build all XML for `add_action_rule`
Going forward I plan to make all SOAP request builders build the entire XML, even if that prevents some valid requests from being constructed. To be clear the goal is to allow all valid requests from being built using the fluent API, but support for additional requests will be added as needed. Users that require more flexibility should be able to easily fall back on the APIs for parsing responses, or even `SimpleRequest` in this case, until the ergonomic API has been updated. The reason for this is that the biggest value add comes from these request builders is building the XML while adhering to all the namespace rules expected by the server.