feat(vapix)!: Build all XML for `add_action_rule` (#38)
The most difficult thing with using the SOAP style APIs, after properly
parsing error responses, is crafting an XML request that the server will
accept. As such this API does not make much sense if the user has to
craft the part of the request themselves.

One downside is that the new API probably does not support all valid
requests that a user may want to send, but considering that I mean for
the crate to also provide an API for robust response handling I think
falling back on that should be an acceptable fallback.

---

`crates/vapix/src/services/action1/action_rules.rs`:
- Build the XML manually because this is easier in the short term than
creating a type that serializes correctly, since the latter still
requires a lot of trial and error for me. I am also suspicious of the
storage footprint of serde trait implementations and the traits feel
even more like a leaky abstraction when working with xml.
- Accept the topic expression and message content as strings because I
don't have a good guess for how these would be further decomposed. It
would however be nice to provide a builder API for them, too.

`crates/vapix/tests/serde.rs`:
- Update the serde test to create a valid request that exercises more of
the API. We know the request is valid because the same request is
constructed in the smoke tests.