feat: Add deactivate command (#42)
After I started configuring the active device using the filesystem the
smoke tests fail until I either replace it with an accessible device, or
clear it manually. This is annoying and can easily be solved with a
simple, sensible command.

---

`crates/dut/src/lib.rs`:
- Clearing `AXIS_DEVICE_IP` is sufficient for `from_env` to return
`None` and by not clearing more information, we don't unnecessarily
interfere with environment variables that a user may have set up
manually. To be clear, this is only a hypothetical benefit but that is
the motivation for not including the whole list.