fix: Use correct program name in completions help (#43)
`crates/bin-utils/src/completions_command.rs`:
- I investigated modifying the command created by `#[derive(Parser)]` to
make it easier keeping all the help texts up to date, varying only the
program name, but I couldn't quickly find a way to parse it back into
the `Cli` struct. Because it makes sense that this would not be a
supported use case, I didn't spend much time on this.