feat: Emit warnings by default and save full log to file (#44)
Prompted by me using the export command and being confused about why I
didn't have a usable device exported.

Since I'm taking a look at the logging I thought I might as well write
it to file as well; I find that I often don't want to see the logs
unless something actually went wrong, so we write them to file. Unlike
previous implementations I have made of this, we always write the logs
to file, not only when we are not writing them to stderr because it is
often useful to emit warnings and errors to the terminal to make the
user aware of problems.

---

`bin/device-inventory-smoke-test.sh`:
- Specify which device to export because otherwise a warning will be
logged and the logs contain timestamps that are not reproducible.