s: split rest
I know REST is always HTTP, but I don't know what else to call this flavor of API right now. The idea is to separate the parsing of the response, which would likely be reused for other transports. This is also why the request builder is in the http module; the path likely makes it HTTP specific.