feat(vapix)!: Add `ssh_1` (#49)
---

`crates/vapix/src/config/ssh_1.rs`:
- I don't need the `add_user` API, but having it makes it easier to test
the `set_user` bindings automatically.
- The verbs are chosen to match the operation names defined in the
device configuration API to make it easier to consistently map the rust
API to the documentation and to the API model, even though this is not
always the approach taken by the documentation [^1].

`crates/vapix/src/rest.rs`:
- This file retains only the functions that operate on the body of the
responses because I want to make sure that this is reusable
independently of how the request was constructed and sent.

`crates/vapix/src/rest_http.rs`:
- The `rest` and `rest_http` names are chosen mostly to fit the pattern
established for JSON and SOAP and I don't think they are particularly
good, not least because REST is typically (always?) implemented on top
of HTTP.

`crates/vapix/src/config/ssh_1_examples/*_response.json`:
- These are all collected from 11.11.73

[^1]:
https://developer.axis.com/vapix/device-configuration/ssh-management/#modify-an-existing-ssh-user