feat(di): Add option to login using MFA (#51)
The purpose is primarily to make `device-inventory` easier to use, in
particular for users who don't know how to extract a cookie from a
browser and to shorten getting-started instructions for projects that
use it.

I think I was also hoping that this would make it more ergonomic for
myself and facilitate use of VLT from CI, but I hadn't really reasoned
through either and both need more work:
- For me, having to type my email and copy-paste my password to the
terminal is about the same amount of work as using the browser to
autofill, then copy-pasting the cookie.
- For CI, there still needs to be a way to handle MFA. Perhaps it's
possible to implement TOTP in Rust and store the long-term secret in CI.