feat(vlt): Add library crate (#52)
This commit factors out the VLT code to a separate library crate and
implements support for starting and canceling loans.

As was increasing the coverage and quality of the bindings to implement
the ensure command in device-inventory, it made more sense improve the
separation and to make it reusable.

---

`crates/vlt/src/authentication.rs`:
- Add the option to create session ids to make it possible for users to
obtain these without going through the authentication flow.

`crates/vlt/src/client.rs`:
- Add a custom client type to make it more difficult to forget to set
the authentication header.

`crates/vlt/src/requests.rs`:
- Use functions to create request builders to be consistent with how I
have been building the VAPIX API. However, I do think it makes less
sense here since the API surface is smaller which makes it more feasible
to spend more time building out the builder for each endpoint. Having a
builder for each request would allow having multiple constructors, which
could come in handy if some of the arguments that are now assumed to be
required become, or turn out to be, optional.

`crates/vlt/src/responses.rs`:
- Require that responses implement serialize in order to enable the
lossless assertion. The reason for having this assertion is primarily to
allow this crate to act as documentation for the APIs, which are not
officially documented, one of the main benefits of documentation is to
show what information is available, and by ensuring the deserialization
captures all information we can also ensure that the documentation is up
to date.

`crates/vlt/tests/responses/post_loans_days.json`:
- These are curated and edited to document or test various things. The
devices list is especially long, typically over 4k lines, most of which
is not helpful.