feat(vlt): Add `Loan.host`
It makes sense that VLT provide this information, especially to help users avoid trying to use `external_ip`; I don't know what the external- and internal-IP addresses are, but they are not the IP advertised in the GUI which is what we can actually use to connect to the device.

Also add some documentation to further help users avoid this mistake.