feat(vlt): Add `host` method to `Loan` (#54)
It makes sense that VLT provide this information, especially to help
users avoid trying to use `external_ip`; I don't know what the external-
and internal-IP addresses are, but they are not the IP advertised in the
GUI which is what we can actually use to connect to the device.

---

`crates/vlt/src/responses.rs`:
- Docstrings are added to help users, including myself, avoid
accidentally using the `external_ip` incorrectly.
- The new method returns a host to future proof in case VLT moves to
IPv6 or a domain; I generally want consumers to be agnostic to what kind
of host it is anyway.