feat(di): Add status column to list command output
This is helpful for checking which device will be used when running other commands, and for inspecting the effect of other commands.

Prompted by me adding an abandon command which definitely should deactivate the device, but I had forgotten to add it and was unsure if I needed to or if it was already implied by remove.

Note that for better or worse, removing a device does not deactivate it. If remove ends up being more of a plumbing command, then perhaps it makes sense. If not, then it's probably surprising. In any case leave the current behavior intact until I have a stronger opinion.