feat(di): Add status column to list command output (#56)
This is helpful for checking which device will be used when running
other commands, and for inspecting the effect of other commands.

Prompted by me adding an abandon command which definitely should
deactivate the device, but I had forgotten to add it and was unsure if I
needed to or if it was already implied by remove.

Note that for better or worse, removing a device does not deactivate it.
If remove ends up being more of a plumbing command, then perhaps it
makes sense. If not, then it's probably surprising. In any case leave
the current behavior intact until I have a stronger opinion.

`bin/device-inventory-smoke-test.sh`:
- Consolidate the list commands to make the output easier to compare and
correlate with the commands that changed it.
- Activate a device using environment variables before listing to make
sure the output is reproducible in development where there may be an
active device configured on the filesystem.

`crates/device-inventory/src/commands/list.rs`:
- The ad-hoc fingerprinting in two places is ugly, but it works and I
would rather take on a little bit of tech debt now and then figure out
the device data model later when there is a bit more code to infer
patterns from.
- No two correctly configured devices will have the same IP and HTTP
port, or traffic could not be routed. There could of course be problems
with obsolete information in the database so maybe this is a heuristic I
will need to reconsider.