fix(vlt)!: Parse external IP eagerly
Prompted by me wanting to use it in a callback where it would not be possible to propagate the error using `?`.

As far as I can tell, the external IP is always an IPv4 address, and by enforcing this when the responses are parsed using the responses becomes more ergonomic.

---

`crates/vlt/src/responses.rs`:
- Add a custom `ExternalIp` type to emphasize that this value is probably not useful as an IP address.