fix(vlt)!: Parse external IP eagerly (#57)
Prompted by me wanting to use it in a callback where it would not be
possible to propagate the error using `?`.

As far as I can tell, the external IP is always an IPv4 address, and by
enforcing this when the responses are parsed using the responses becomes
more ergonomic.

---

`crates/vlt/src/responses.rs`:
- Add a custom `ExternalIp` type to emphasize that this value is
probably not useful as an IP address.
- Change `assert_eq!` to `debug_assert_eq!` because I want the program
to continue working for users if this assumption breaks. The program
does not depend on the two being equivalent, only on the one returned
being correct, and the returned version is the one that is used by the
reference implementation. The base port implementation was inferred from
examples.