chore(v,vlt): Log response text before parsing (#58)
Consistently log the response text before parsing so that when a request
fails to parse, I don't have to reproduce the response before I can
troubleshoot the parsing failure.

---

`crates/device-inventory/src/db.rs`:
- The same treatment is not given to file reads, because these should be
easier to reproduce.

`crates/vapix/src/soap_http.rs`:
- Use `trace!` instead of `println!` to avoid polluting the terminal
when it works; the logs will typically be written to a file that can be
consulted after a failure anyway.

`crates/vlt/src/authentication.rs`:
- Status is not logged mainly to not increase the complexity of the code
for the sake of debug logging. I also expect it to be less useful here
because `error_for_status` decreases the number of different statuses
that it could possibly be.

`*`:
- Don't log the responses in release builds to mitigate the security
implications of logging the full response from the server. The responses
typically don't contain highly sensitive information, but it is still
good to be careful.
- Leave texts in error messages in place for now because I clearly
thought they would be useful. When I next encounter a parsing failure, I
should evaluate which method of capturing the errors is more useful.
- Use `cfg!(debug_assertions)` instead of `#[cfg(debug_assertions)]`
mainly because it avoids issues like unused imports when the release
profile is used. This is acceptable because the speed or size of these
dev tools is not critical, and with a bit of luck the compiler will
prune these branches.