feat(di): list devices from multiple sources (#61)
This is the first command to be rewritten using an interface model where
`device-inventory` is no longer an interface primarily to the local
database, but to multiple sources. This is driven by me wanting to
automate the workflow of borrowing, adding, and activating a device. My
hypothesis is that all the commands can be rewritten with a more
declarative API where:
- the command describes a desired state (arguments will probably be
added to control factory reset, initialization, etc),
- a set of filter arguments describe which device(s) what device it is
the user wants to work with, and
- a set of sources describe from where the program is allowed to source
the device.

---

`bin/device-inventory-smoke-test.sh`:
- Specify the `dut` and `di` sources because other sources depend on
state that is external to the machine on which the tests run.

`crates/device-inventory/src/commands/list.rs`:
- Factor out the table to make it easier to check that all columns are
pushed to for every device. This is messy and can definitely be improved
further, but I don't want to spend time figuring out if I should do it
myself with `itertools`, use an off-the-shelf table printing library, or
something else before I know the functionality is definitely staying.
- The priority column replaces the old _active_ column and is mainly
intended to help me troubleshoot the list command and commands that
change the state of the sources.
- Add `probe` option primarily to make it possible to filter devices
from dut and inventory by model. It is opt-in for now because I prefer
to not make fallible calls to the network (perhaps the vlt and mdns
sources should be disabled by default for the same reason...)
- `probe_devices` runs before VLT devices are added to avoid hammering
the whole VLT inventory with requests. Probing devices on-loan to the
current user would make sense, but is deferred because it would increase
the code complexity.

`crates/device-inventory/src/mdns_source.rs`:
- Copy code from `device-finder` because it is not a lot and it is too
early to conclude if tying the implementations together with shared code
makes sense.