fix(df,di): Discover services in parallel
Now that this happens as part of `device-inventory list` I find it annoying that the discovery takes so long time. This change decreases the time it takes from about 7s to about 3s when build in dev mode and run with sources di and mdns.

I'm slightly surprised the speedup is closer to 3x since with `--source di` only, the program runs in a fraction of a second, but the improvement is meaningful and I don't want to go back down the mDNS rabbit hole at this time.