feat(v): Try get brand API when determining scheme
The system ready API was introduced in AXIS OS 9.50, so for earlier versions `build_with_automatic_scheme` would fail. Such old devices aren't the focus of these tools since they are probably the target of only a small portion of development, but creating a client is the foundation for testing every other API and support can be added without hurting other use cases much (detecting the scheme probably takes a bit longer, but mostly for http and this should not be on a hot path anyway).

Note that it still fails if the root user has been set up and the authentication has not been changed to basic. Since this is the state that VLT devices are in when checked out, this problem will be encountered frequently. But at least there is a workaround in enabling basic authentication.