feat(di): Replace the sources concept
I realized when using the old version that I like how snappy it is and how concise the output is. The same can be accomplished with the sources, but specifying `--source dut --source di` is tedious and adding options to exclude the remote sources would add complexity and be equally tedious.

---

`crates/device-inventory/src/commands/list.rs`:
- Replace `--source` with two separate flags to make it easier to type.
- Always include the local sources because the user has already decided that they are in some way relevant.