chore(di)!: Replace json import (#71)
This will make it easier to add additional attributes to saved devices,
such as architecture, since we don't have to support adding devices
inferred from loans without the additional attributes. The need for this
functionality has been reduced since the login command made it easier to
set the cookie needed for authentication and there is now test coverage
for the parsing of loan responses in the `rs4a-vlt` crate.

---

`bin/device-inventory-smoke-test.sh`:
- Overwrite the `devices.json` snapshot to make it easier to detect if
the snapshot becomes outdated.

`crates/device-inventory/src/commands/dump.rs`:
- Added to make it easier to seed the database in tests.
- Named `dump` instead of `export` because calling it's twin `import`
would suggest it adds to the state whereas `load` suggests the state is
replaced, which is the behavior I want.

`crates/device-inventory/src/commands/load.rs`:
- Added to make it easier to test the binary relying only on public
interfaces, and to make it easier to switch between inventories by
dumping and loading them.
- The input function is moved from the input command because I don't
have a reason to re-evaluate the input method at this time.