feat(di): Filter on firmware (#72)
I realized while working on #64 that this is an important parameter when
selecting a device to work on, not least because the tooling provided by
this project has poor support for devices with a firmware older than 11.

---

`*`:
- Use the term firmware to stay clear of trademarks and because that is
the term used by the VLT.

`crates/device-inventory/src/fusion.rs`:
- Use version requirement as filter because it is a proven and well
known way to specify versions that is generally better aligned with what
we want to select than a glob pattern would be.
- The coercion to SemVer is needed for older firmware which uses a
versioning scheme that is similar, but not strictly the same. I am not
completely sure in what ways it can vary, but from what I have observed
the versions sometimes have a fourth segment. I'm betting on this rarely
being relevant for firmware selection, but to make it show up in the
output it is parsed as build metadata.
- Add `dedup` because without it any device fused from more than one
source would panic.
- Parse firmware version when adding information to the fused device
instead of when using the firmware version because it is easier to deal
with the error. Ideally the parsing would be pushed into the response
parsing of VAPIX and VLT, but I'm not confident that it will never fail
so I don't want to push it into the libraries.
- Remove unused `From` implementations because I realized they are
unused and it is not immediately clear how to return an `&Version` from
a `String`.