feat: Probe for architecture (#73)
I also considered adding the architecture eagerly when a device is added
to the inventory, but the current implementations has some advantages:
- Works for VLT loans (which provide credentials, but no architecture)
- Works the same for hard (e.g. architecture), firm (e.g. firmware) and
soft (e.g. uptime) attributes.
- Devices can be saved without access (niche use case).

Drawbacks compared to eager probing include:
- The cost of probing the device is paid every many times instead of
only once. But since firm and soft attributes need to be queried lazily,
it is probably rare that the lazy implementation will issue a query when
an eager implementation would not.
- It is easy to forget to set the `--probe` flag, and not doing so would
when e.g. the architecture filter is used would cause saved devices to
not show up in the list.

---

`crates/device-inventory/src/fusion.rs`:
- Replace `replace_vlt_device` with `add_vlt_device` because it is
convenient to have a single method replace both the vlt device and the
architecture. Other methods are not replaced because I am not confident
in this pattern yet and I don't want to do unnecessary work.