feat(v): Make `reqwest/rustls-tls` optional (#76)
Prompted by me experimenting with using OpenSSL in an ACAP app where
dynamically linking OpenSSL could greatly reduce the storage used. The
app depends on `rs4a-vapix`, which means it could not completely get rid
of the transient dependency on `rustls`.

Since the other crates are binaries intended to run on host and I don't
have a reason to build them with OpenSSL, `rustls` is left as the only
option.